Fix default sqlite conversation to long

+++ src/Financier.Common/Converters/CategoryTitleConverter.cs
@@ -11,7 +11,7 @@ namespace Financier.Common.Converters
         public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
         {
             var title = (string)values[0];
-            var level =  (int)(long)values[1];
+            var level =  (int)values[1];
             return (title ?? string.Empty).PadLeft((title ?? string.Empty).Length + level, '-');
         }
