+++ src/Financier.Reports/Dynamic/ReportDynamicDebitCretitPayeeVM.cs
@@ -96,7 +96,7 @@ order by
 
             foreach (var item in list.OrderBy(x => x.Year).ThenBy(x => x.Month))
             {
-                values.Points.Add(DateTimeAxis.CreateDataPoint(new DateTime((int)item.Year, (int)item.Month, 1), item.Total ?? 0));
+                values.Points.Add(DateTimeAxis.CreateDataPoint(new DateTime(item.Year, item.Month, 1), item.Total ?? 0));
             }
 
             model.Series.Add(values);
