+++ src/Financier.Reports/Dynamic/ReportDynamicRestVM.cs
@@ -94,7 +94,7 @@ ORDER BY year, month, day";
 
             foreach (var item in list.OrderBy(x => x.Year).ThenBy(x => x.Month).ThenBy(x => x.Day))
             {
-                lineSeries.Points.Add(DateTimeAxis.CreateDataPoint(new DateTime((int)item.Year, (int)item.Month, (int)item.Day), item.Total ?? 0));
+                lineSeries.Points.Add(DateTimeAxis.CreateDataPoint(new DateTime(item.Year, item.Month, item.Day), item.Total ?? 0));
             }
 
             model.Axes.Add(dateTimeAxis);
